Remove chirp and _conn from ch_message_t
The message may not have any direct association with chirp or a
connection, since it can be used with different chirp instances (though
not at the same time).

There is an direct association with a chirp instance through the
buffer-pool if the message created by a chirp instances. This allow the
user to receive a message with one chirp instance and sending it with
another (routers).

When releasing the such a message, it has to be released on the right
buffer, therefore the _pool pointer.