Tutorial: A simple monitoring solution
Introduce a tutorial for Chirp, to show users roughly how it works. The
tutorial introduces the user via a simple monitoring solution that
consist of the following parts that communicate with each other:

* An "agent" program that periodically checks a TCP service for its
  availability, and reports it via chirp to an upstream node
* A "monitor" program that collects and displays the service's status
* A "collector" program that collects data from multiple agents and
  forwards it to another collector, or a monitor.

The monitor checks upon the service list periodically and notes "agent
down" if it didn't hear from the agent for some time (twice the reported
polling interval).