Test and implement one handler
This pull-request is all about doing things in the right order, we
allowed the following to happen:

```
   client                    echo-server

         --------M1----------> handlers++
        <-------ack----------
         --------M2----------> stop stream
        <--------M1-echo-----
         -------ack----------> deadlock
```

The key to solve this is to move sending the ack to
ch_chirp_release_message: we only want to send the ack when we release a
handler or we run out of handler.

Additionally:

* We check for available handlers when we need them, not in advance
  since the handler could have been released in the mean time

* We do not use a handler for ack message, since no two ack message can
  be active, it is ok to use memory owned by the connection.

I also remove all retry related code, because I decided to drop that
feature.