Document new findings about chirp
Something dramatic has changed with the last PR: config.ACKNOWLEDGE=1 is
connection-synchronous. Its not possible to have limited resources,
acknowledge and be asynchronous (deadlocks). The python prototype used
dynamic resources and didn't have that problem. But I thought about it a
lot and this is much more predicable and problems in the python
prototype disappear: we don't need retry and flow-control.